requirements.txt
asgiref==3.8.1
black==24.8.0
bleach==6.1.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
gunicorn==23.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.5
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
webencodings==0.5.1
